+++ firmware/CHANGELOG.md
@@ -41,6 +41,7 @@ or
  - Add 1-5-4-8-3-7-2-6 and 1-6-5-10-2-7-3-8-4-9 firing orders
  - Console command `set_sensor_mock` now accepts a sensor by name, instead of index. Example: `set_sensor_mock CLT 85.5`
  - TunerStudio UI improvements (#436, etc)
+ - Automatic calculation of knock sensing frequency based on cylinder bore #379
 
 ### Fixed
  - Improve performance with Lua CAN reception of a high volume of frames